Add dialog to add price
To add a new price on a beer, there is now a dialog to input a new price.
There is 2 news routes (for add price, and edit beer).
The new price is saved in the db.

Just now, you can just add a new price. You can't add a new beeeeeers. Sorry.